refactor(security): extract feesOnly logic into dedicated classes
